Added test structure, dummy methods + refactored 'Portofolio'->'Portfolio'
