+++ src/main/java/org/ek/portfoliobackend/service/test-dummy:slettes når vi har filer i package
