+++ src/test/java/org/ek/portfoliobackend/controller/ProjectControllerTest.java
@@ -1,0 +1,23 @@
+package org.ek.portfoliobackend.controller;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.webmvc.test.autoconfigure.AutoConfigureMockMvc;
+import org.springframework.test.web.servlet.MockMvc;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+@SpringBootTest
+@AutoConfigureMockMvc
+public class ProjectControllerTest {
+
+    @Autowired
+    private MockMvc mockMvc;
+
+    //dummy method for test-setup
+    @Test
+    void mockMvcIsConfigured() {
+        assertThat(mockMvc).isNotNull();
+    }
+}
