+++ src/test/java/org/ek/portfoliobackend/service/dummy-file
@@ -1,0 +1,1 @@
+created to hold structure, delete when classes are added.
