+++ src/test/java/org/ek/portfoliobackend/service/impl/ProjectServiceTest.java
@@ -1,0 +1,16 @@
+package org.ek.portfoliobackend.service.impl;
+
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.junit.jupiter.MockitoExtension;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+@ExtendWith(MockitoExtension.class)
+public class ProjectServiceTest {
+
+    @Test
+    void testFrameworkWorks(){
+        assertThat(true).isTrue();
+    }
+}
